preprocess.randomize: move tests to test_randomize
